Version 1.1.5 - Revert to desktop only & use adapter API where available
The Adapter API is relative to the open vault and does not support absolute
file paths. This prevents the plugin from reading the nickname file in
other vaults, particularly when they are on a different drive than the
active vault. Therefore, the plugin still uses `fs` member of the adpater
for reading vault nicknames between vaults.