Fix #1: Use custom menus for the vault switcher's click and contextmenu events
This allows the nicknames to display correctly on Mac. Previously, macOS's
context menu failed to be hooked because it was rendered using the OS's
native UI and therefore wasn't part of the DOM.

The custom menu added in this commit is simpler and more flexible. (We now
show an icon for the 'Vault Nickname settings' context menu item.)

The caveat is that Obsidian may update/add features to their default vault
switcher menus in future updates. To account for missing features in this
plugin's recreated menus, holding the Shift key when clicking the vault
switcher menu invokes the original, unaltered menus.